requirements.txt
tabula-py
